Moving away from tiling system - part 1 - IntRect -> FRect

+++ game/src/gameui/game_hud.rs
@@ -6,7 +6,7 @@ pub fn update_game_hud(context: &mut GameContext) {
     context.basic_info_hud.update();
 }
 
-pub fn hud_ui(context: &GameContext, width: i32, height: i32, show_debug_info: bool, fps: u32) -> Layout {
+pub fn hud_ui(context: &GameContext, width: f32, height: f32, show_debug_info: bool, fps: u32) -> Layout {
     let is_dead = context.is_dead();
     let game_mode = &current_game_mode();
