+++ game_core/src/equipment/melee.rs
@@ -63,19 +63,19 @@ impl Entity {
     } 
 }
 
-fn bullet_offsets(direction: Direction) -> Vec<(i32, i32)> {
+fn bullet_offsets(direction: Direction) -> Vec<(f32, f32)> {
     match direction {
         Direction::Up => vec![
-            (-1, -1), (0, -2), (0, -1), (1, -1)
+            (-1.0, -1.0), (0.0, -2.0), (0.0, -1.0), (1.0, -1.0)
         ],
         Direction::Down | Direction::Unknown | Direction::Still => vec![
-            (-1, 1), (0, 2), (0, 1), (1, 1)
+            (-1.0, 1.0), (0.0, 2.0), (0.0, 1.0), (1.0, 1.0)
         ],
         Direction::Right => vec![
-            (1, -1), (2, 0), (1, 0), (1, 1)
+            (1.0, -1.0), (2.0, 0.0), (1.0, 0.0), (1.0, 1.0)
         ],
         Direction::Left => vec![
-            (-1, -1), (-2, 0), (-1, 0), (-1, 1)
+            (-1.0, -1.0), (-2.0, 0.0), (-1.0, 0.0), (-1.0, 1.0)
         ],
     }
 }
