+++ game_core/src/prefabs/all.rs
@@ -1,6 +1,6 @@
 use crate::{entities::species::Species, features::entity::Entity, prefabs::{house_single_floor::new_house_single_floor, house_small::new_small_house, house_two_floors::new_house_two_floors, shop::new_shop}};
 
-pub fn new_building(source_world_id: u32, x: i32, y: i32, species: &Species) -> Vec<Entity> {
+pub fn new_building(source_world_id: u32, x: f32, y: f32, species: &Species) -> Vec<Entity> {
     if is_small_house(species.id) {
         return new_small_house(species, source_world_id, x, y);
     }
